babel-echo/requirements.txt
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-speech==2.0.0
google-cloud-texttospeech==2.2.0
google-cloud-translate==3.0.1
googleapis-common-protos==1.52.0
grpcio==1.33.2
idna==2.10
libcst==0.3.13
mypy-extensions==0.4.3
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pygame==2.0.0
python-vlc==3.0.11115
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
